requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
reportlab==4.0.7
python-dotenv==1.0.0
requests==2.31.0
